Handles existing client during registration
Improves client registration by checking for existing clients based on CPF.
If a client with the provided CPF already exists, it returns the existing client, preventing duplicate entries.
Also, updates the client store method in the controller to return HTTP 200 with success message for existing client.

Also, restarts policy in docker-compose are changed to `no` or `unless-stopped`.